inCaousel eager loading fixes siziing issues, also reverted to hardcoded address to BFF call because it was not working in FLY. Ill talk to Will about that at some point
